Fixed bug: [1072439] Crystalfontz 632 - not working at all. This was because flow control was disabled. Sadly I was unable to enabled in at run-time, so was forced to do it at design time (using two serial objects - one with flow control enabled and the other not).
Tidied up some of the LCD serial code.
Had to disable some MO probing code - it needs to live in ULCD_MO but I couldn't understand what it was doing, well enough, to do the move.  Will wait for a user to complain and hopefully they will explain what they used to see in the setup screen :-)