Update version to 5.3
Fix bug where scrolling (and other things) would stop working at high refresh rates.

It seems that all timers except the refresh timer were being ignored. Removing the setting of the interval to 0 fixed the problem. It could be that when a timer is reset, it is placed at the front of the queue, as the refresh timer was being done every 12ms - it was always at the front.

I also made some other changes (such as renaming timer1 to timerRefresh) whilst I was trying to track this bug down.